MAINT: Move conversion functions to lti_conversion
